pyproject.toml
[project]
name = "amazon-search-retrieval"
version = "0.1.0"
description = "ESCI cross-encoder reranker for product search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "jupyter>=1.1.1",
    "ipykernel>=6.0.0",
    "datasets>=4.6.1",
    "accelerate>=1.12.0",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
